feat : 기본 레이아웃 구성 및  router 작업 완료
