+++ src/pages/Home.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const Home = () => {
+  return (
+    <div>
+      <div>Home</div>
+    </div>
+  );
+};
+
+export default Home;
