--- src/pages/Typography.tsx
@@ -1,26 +1,0 @@
-import React from 'react';
-
-const Typography = () => {
-  return (
-    <div>
-      <div>
-        <p className="text-28-bold m-2">text-28-bold</p>
-        <p className="text-24-bold m-2">text-24-bold</p>
-        <p className="text-24-regular m-2">text-24-bold</p>
-        <p className="text-20-bold m-2">text-20-bold</p>
-        <p className="text-20-regular m-2">text-20-regular</p>
-        <p className="text-18-bold m-2">text-18-bold</p>
-        <p className="text-18-regular m-2">text-18-regular</p>
-        <p className="text-16-bold m-2">text-16-bold</p>
-        <p className="text-16-regular m-2">text-16-regular</p>
-        <p className="text-15-bold m-2">text-15-bold</p>
-        <p className="text-15-regular m-2">text-15-regular</p>
-        <p className="text-14-bold m-2">text-14-bold</p>
-        <p className="text-14-regular m-2">text-14-regular</p>
-        <p className="text-12-regular m-2">text-12-regular</p>
-      </div>
-    </div>
-  );
-};
-
-export default Typography;
